update code reception for batch

+++ lib/src/presentation/providers/db/picking/tbl_pick_products/pick_products_table.dart
@@ -72,7 +72,7 @@ class PickProductsTable {
         $columnOrigin VARCHAR(255),
         $columnObservation TEXT,
         $columnUnidades TEXT,
-        $columnWeight INTEGER,
+        $columnWeight REAL,
         $columnIsMuelle INTEGER,
         $columnMuelleId INTEGER,
         $columnIsLocationIsOk INTEGER,
